refactor(project): separar contexto en projectsCtx y corregir efecto en useProjects
- Extrae ProjectsContext y useProjectsContext a projectsCtx.ts para
  cumplir la regla de Fast Refresh (los .tsx solo exportan componentes)
- Reescribe el efecto de carga en useProjects usando .then() para evitar
  setState en el cuerpo síncrono del efecto (React Compiler)
- Elimina setLoading(true) redundante (useState ya inicializa en true)
- Actualiza imports en ProjectSidebarContent, ProjectsPage e index.ts
