+++ src/features/project/components/ProjectsPage.tsx
@@ -2,7 +2,7 @@ import { useState } from "react"
 import { IconPlus } from "@tabler/icons-react"
 import { SidebarTrigger } from "@/shared"
 import { Button } from "@/shared"
-import { useProjectsContext } from "../context/ProjectsContext"
+import { useProjectsContext } from "../context/projectsCtx"
 import { ProjectCard } from "./ProjectCard"
 import { CreateProjectModal } from "./CreateProjectModal"
 import type { ProjectFormValues } from "../schemas/project.schema"
